Inline people view query hooks
