+++ apps/app/src/__tests__/authenticated-routes-source.test.ts
@@ -355,7 +355,7 @@ describe("app authenticated route migration", () => {
     const searchSource = source("src/people/people-search-params.ts");
     const querySource = source("src/people/use-people-list-query.ts");
     const viewsSource = source("src/people/people-view-switcher.tsx");
-    const viewQuerySource = source("src/people/use-people-views-query.ts");
+    const viewQuerySource = viewsSource;
     const toolbarSource = source("src/people/people-toolbar.tsx");
     const tableSource = source("src/people/people-table-view.tsx");
     const detailSource = source("src/people/people-detail-content.tsx");
